BUG-2663: Support v6 in BGP flowspec
Yang model:
- separated ipv4 and ipv6 distinct component types
- removed component type
Created abstract parser with common functionality and two
    parsers with specific functionality
Updated RIB support for ipv4, created new for ipv6 and moved
    common functionality into abstract RIB support class
Registred parsers and support for afi-s
Updated, expanded and created tests for parsers.

+++ util/src/main/java/org/opendaylight/protocol/util/Values.java
@@ -29,7 +29,7 @@ public final class Values {
     /**
      * Maximum unsigned Integer value (2147483648);
      */
-    public static final long UNSIGNED_INT_MAX_VALUE = Integer.MAX_VALUE + 1;
+    public static final long UNSIGNED_INT_MAX_VALUE = (long) Integer.MAX_VALUE + 1;
 
     /**
      * Maximum unsigned Byte value in hex (0xFF).
